+++ src/cc-context/refs.js
@@ -1,0 +1,4 @@
+
+const refs = {};
+
+export default refs;