+++ src/core/base/catch-cc-error.js
@@ -1,6 +1,7 @@
-import ccContext from '../../cc-context';
+import runtimeHandler from '../../cc-context/runtime-handler';
 
-export default err=>{
-  if(ccContext.errorHandler)ccContext.errorHandler(err);
+export default err => {
+  const errorHandler = runtimeHandler.errorHandler;
+  if (errorHandler) errorHandler(err);
   else throw err;
 }